test(web): move MCP page objects into tests/e2e/pages/
Introduce a central tests/e2e/pages/ directory for page objects and move the
four POM classes (AdminMcpServersPage, ChatPreferencesPage, AgentEditorPage,
ActionsPopover) there. McpOAuthFlow (flow glue) and mcpToolInvocation (helper
functions) stay in mcp/ since they are not page objects. Document pages/ as the
standard home in tests/e2e/README.md; existing co-located POMs migrate later.
