+++ web/tests/e2e/pages/AdminMcpServersPage.ts
