+++ web/tests/e2e/pages/ChatPreferencesPage.ts
